lib/pyproject.toml
[project]
name = "continuum"
version = "0.1.0"
description = "Continuum Python SDK"
authors = [{ name = "Antonio Zugaldia" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.12"

# Dependencies are pinned to minor versions (~=) to only allow automatic patch updates
dependencies = [
    "beautifulsoup4~=4.14.3",
    "elevenlabs~=2.27.0",
    "faster-whisper~=1.2.1",
    "google-genai~=1.56.0",
    "kokoro~=0.9.4",
    "mistune~=3.2.0",
    "ollama~=0.6.1",
    "openai~=2.14.0",
    "pydantic~=2.12.5",
    "pydantic-ai~=1.39.0", # Switch to the pydantic-ai-slim [] format
    "reachy-mini[gstreamer]~=1.2.6",
    "requests~=2.32.5",
    "roslibpy~=2.0.0",
    "typer~=0.21.0",
    "websocket-client~=1.9.0",
]

[project.optional-dependencies]
audio = [
    "librosa~=0.11.0"
]
mic = [
    "continuum[audio]",
    "pvrecorder~=1.2.7",
    "PyAudio~=0.2.14"
]
vad = [
    "continuum[audio]",
    "silero-vad~=6.2.0",
    "pvcobra~=3.0.1",
]

[project.scripts]
continuum = "continuum.continuum:cli"

[tool.poetry]
packages = [{ include = "continuum", from = "src" }]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
